feat: enable `-Wsign-compare`
This enables the `-Wsign-compare` warning everywhere except inside the
`ASSERT_EQ` macro, which will be fixed in later commits.

+++ src/arch/aarch64/hypervisor/ffa.c
@@ -44,7 +44,7 @@ void arch_ffa_init(void)
 
 	if (ret.func == FFA_SUCCESS_32) {
 		spmc_id = ret.arg2;
-	} else if (ret.func == SMCCC_ERROR_UNKNOWN ||
+	} else if (ret.func == (uint64_t)SMCCC_ERROR_UNKNOWN ||
 		   (ret.func == FFA_ERROR_32 &&
 		    ffa_error_code(ret) == FFA_NOT_SUPPORTED)) {
 		spmc_id = HF_SPMC_VM_ID;
